+++ sumo/src/gui/GUITLLogicPhasesTrackerWindow.cpp
@@ -173,8 +173,8 @@ GUITLLogicPhasesTrackerWindow::GUITLLogicPhasesTrackerWindow(
     myConnector = new GLObjectValuePassConnector<std::pair<SUMOTime, MSPhaseDefinition> >(wrapper, src, this);
     FXint height = (FXint)(myTLLogic->getLinks().size() * 20 + 30 + 8 + 30);
     app.addChild(this, true);
-    for (size_t i = 0; i < myTLLogic->getLinks().size(); ++i) {
-        myLinkNames.push_back(toString<size_t>(i));
+    for (int i = 0; i < myTLLogic->getLinks().size(); ++i) {
+        myLinkNames.push_back(toString<int>(i));
     }
     FXVerticalFrame* glcanvasFrame =
         new FXVerticalFrame(this,
@@ -200,8 +200,8 @@ GUITLLogicPhasesTrackerWindow::GUITLLogicPhasesTrackerWindow(
     FXint height = (FXint)(myTLLogic->getLinks().size() * 20 + 30 + 8);
     setTitle("TLS-Tracker");
     app.addChild(this, true);
-    for (size_t i = 0; i < myTLLogic->getLinks().size(); ++i) {
-        myLinkNames.push_back(toString<size_t>(i));
+    for (int i = 0; i < myTLLogic->getLinks().size(); ++i) {
+        myLinkNames.push_back(toString<int>(i));
     }
     FXVerticalFrame* glcanvasFrame =
         new FXVerticalFrame(this,
@@ -240,7 +240,7 @@ GUITLLogicPhasesTrackerWindow::drawValues(GUITLLogicPhasesTrackerPanel& caller) 
     // compute what shall be shown (what is visible)
     myFirstPhase2Show = 0;
     myFirstPhaseOffset = 0;
-    size_t leftOffset = 0;
+    int leftOffset = 0;
     myFirstTime2Show = 0;
     if (!myAmInTrackingMode) {
         myPhases.clear();
@@ -266,7 +266,7 @@ GUITLLogicPhasesTrackerWindow::drawValues(GUITLLogicPhasesTrackerPanel& caller) 
         // check whether no phases are known at all
         if (myDurations.size() != 0) {
             SUMOTime durs = 0;
-            size_t phaseOffset = myDurations.size() - 1;
+            int phaseOffset = myDurations.size() - 1;
             DurationsVector::reverse_iterator i = myDurations.rbegin();
             while (i != myDurations.rend()) {
                 if (durs + (*i) > beginOffset) {
@@ -309,7 +309,7 @@ GUITLLogicPhasesTrackerWindow::drawValues(GUITLLogicPhasesTrackerPanel& caller) 
     // draw the link names and the lines dividing them
     SUMOReal h = (SUMOReal)(1.0 - h10);
     SUMOReal h2 = 12;
-    for (size_t i = 0; i < myTLLogic->getLinks().size() + 1; ++i) {
+    for (int i = 0; i < myTLLogic->getLinks().size() + 1; ++i) {
         // draw the bar
         glBegin(GL_LINES);
         glVertex2d(0, h);
@@ -353,12 +353,12 @@ GUITLLogicPhasesTrackerWindow::drawValues(GUITLLogicPhasesTrackerPanel& caller) 
     // and the initial phase information
     PhasesVector::iterator pi = myPhases.begin() + myFirstPhase2Show;
     DurationsVector::iterator pd = myDurations.begin() + myFirstPhase2Show;
-    size_t fpo = myFirstPhaseOffset;
+    int fpo = myFirstPhaseOffset;
 
     // start drawing
-    for (size_t i = 30; pd != myDurations.end();) {
+    for (int i = 30; pd != myDurations.end();) {
         // the first phase may be drawn incompletely
-        size_t duration = *pd - fpo;
+        int duration = *pd - fpo;
         // compute the heigh and the width of the phase
         h = (SUMOReal)(1.0 - h10);
         SUMOReal a = (SUMOReal) duration / width;
